Added search fragment

+++ app/src/main/java/com/example/inbook/domain/response/ListPriceX.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ListPriceX(
     @SerializedName("amountInMicros")
-    var amountInMicros: Int,
+    var amountInMicros: Long,
     @SerializedName("currencyCode")
     var currencyCode: String
 )
