+++ app/src/main/java/com/example/inbook/domain/response/RetailPrice.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class RetailPrice(
     @SerializedName("amountInMicros")
-    var amountInMicros: Int,
+    var amountInMicros: Long,
     @SerializedName("currencyCode")
     var currencyCode: String
 )
